chore: organise vtex actions into folders
