+++ packages/pieces/vtex/src/lib/actions/Product/create-product.ts
@@ -1,11 +1,11 @@
 import { createAction, Property } from "@activepieces/pieces-framework";
-import { Product } from "../common/Product";
-import { Replace } from "../common/types";
+import { Product } from "../../common/Product";
+import { Replace } from "../../common/types";
 
-export const updateProduct = createAction({
-    name: "Update-product",
-    displayName: "Update a Product",
-    description: "Update a product in your catalog",
+export const createProduct = createAction({
+    name: "create-product",
+    displayName: "Create New Product",
+    description: "Create a new product to your catalog",
     props: {
         hostUrl: Property.ShortText({
             displayName: "Host Url",
@@ -20,11 +20,6 @@ export const updateProduct = createAction({
         appToken: Property.SecretText({
             displayName: "App Token",
             description: "VTEX App Token",
-            required: true,
-        }),
-        productId: Property.Number({
-            displayName: "Product ID",
-            description: "Set the product ID",
             required: true,
         }),
         Name: Property.ShortText({
@@ -47,10 +42,6 @@ export const updateProduct = createAction({
             displayName: "Category ID",
             required: true,
         }),
-        DepartmentId: Property.Number({
-            displayName: "DepartmentId",
-            required: true,
-        }),
         LinkId: Property.ShortText({
             displayName: "Link ID",
             required: false,
@@ -59,8 +50,9 @@ export const updateProduct = createAction({
             displayName: "Ref ID",
             required: false,
         }),
-        BrandName: Property.ShortText({
-            displayName: "Brand Name",
+        Id: Property.Number({
+            displayName: "Product ID",
+            description: "Set the product ID",
             required: false,
         }),
         IsVisible: Property.Checkbox({
@@ -102,15 +94,14 @@ export const updateProduct = createAction({
         }),
     },
     async run(context) {
-        const { hostUrl, appKey, appToken, productId } = context.propsValue;
-        const productData: Replace<typeof context.propsValue, { hostUrl?: string; appKey?:string; appToken?: string; productId?: number}> = { ...context.propsValue };
+        const { hostUrl, appKey, appToken } = context.propsValue;
+        const productData: Replace<typeof context.propsValue, { hostUrl?: string; appKey?:string; appToken?: string}> = { ...context.propsValue };
         delete productData.hostUrl;
         delete productData.appKey;
         delete productData.appToken;
-        delete productData.productId;
         
         const product = new Product(hostUrl, appKey, appToken);
 
-        return await product.updateProduct(productId, productData);
+        return await product.createProduct(productData);
     },
 });
